Make it depend on website_portal and overwrite it.
Since the chances that `website_portal` becomes autoinstalled are high,
it's better to depend on it and inhibit it than to warn about conflicts.

The bad news is that this cannot be a plain backport anymore, it's an
adaptation of upstream's v10 version to upstream's v9 version.

The good news is that some modules that use upstream's v9 version will
work out of the box.