[FIX][REF][IMP][website_event_register_free_with_sale] Less requirements for free events.
Before this patch, there was a bug.

1. Get some free tickets for an event:

   ![](https://cloud.githubusercontent.com/assets/973709/20307518/1f0d7684-ab40-11e6-862d-cd3fab5344db.png)

2. Press "Order Now".
3. Leave checkout form blank, and press "Confirm".
4. Odoo required too many fields for free events:

   ![](https://cloud.githubusercontent.com/assets/973709/20307435/b936e868-ab3f-11e6-8056-68142794382f.png)

This refactoring fixes that, because workflow was a bit blurry. Now it reuses more core Odoo facilities, and workflow is more straight-forward and better documented inline.

Also, when you were buying both free and non-free tickets, no confirmation was being sent to the user about the free part. This also is improved.

Signed-off-by: Jairo Llopis <yajo.sk8@gmail.com>