Remove unicode characters from fiile
Sass has some problems when there are several files concatenated and all of them indicate a character encoding. This is what Odoo does, so it is better to remove these characters, mostly when they are used only for copyright purposes and have no legal meaning.

I take the chance to update the copyright and use a bang comment.