[FIX] website_form_builder: restore functionality of required & hidden fields
This functionality was lost in v11 migration. Now it is recovered and tested.

What has been recovered:

- Required fields now actually are required.
- When hiding a required field, you are asked for a default value.

A migration script is added, which removes the `o_required` style for forms that were actually not required. This seems to be a safer approach than leaving them looking as required but not being that, or marking them as required and possibly breaking some workflows in forms built with previous bug.

@Tecnativa TT22901