[FIX] website_form_recaptcha: Allow to work with form editor
The implementation of website_form_recaptcha always checks the
recaptcha validation either the recaptcha template code is in the form
or not. This is problematic now as the form editor allows us to modify
any form, even the main `website_crm` one and this clashes with that
rigid design.

As the recaptcha is always checked we'll inject it allways for every
model enabled for website forms.

This way, the recaptcha will be injected automatically without the need
of defining new extensions to the forms.

Thus the module `website_crm_recaptcha` has no effect and can be
uninstalled. Any similar custom extension won't have any effect as we
dinamically remove any recaptcha code from the form and inject the right
one dinamically after it's loaded.

TT35359