[FIX] website_crm_privacy_policy: Fix editor compatibility
The legal terms acceptance checkbox container had incompatible classes
with the editor that caused a property access error. This prevented
the text or link from being edited. This has been fixed to avoid
incompatibilities with the editor itself and with other modules such as
website_form_require_legal which adds the same functionality of
accepting terms for forms.

TT45834