[IMP] Added website fields to support different reCAPTCHA v2 APIs
To be able to make use of the work of this module for other CAPTCHA APIs compatible with reCAPTCHA v2 (ie. hCaptcha), we added fields to allow modification of the API used by the CAPTCHA widget.
To avoid disrupting current implementations of the module, we however still set default fields that correspond to the Google reCAPTCHA API.